rename PowerUps to PowerUp
